barcodereader/requirements.txt
evdev
paho-mqtt
